+++ Vault/src/utopia/vault/sql/Delete.scala
@@ -3,6 +3,7 @@ package utopia.vault.sql
 import utopia.flow.util.CollectionExtensions._
 import utopia.vault.database.Connection
 import utopia.vault.model.immutable.Table
+import utopia.vault.model.immutable.TableUpdateEvent.DataDeleted
 
 /**
  * This object is used for creating sql statements which delete contents from the database. 
@@ -19,8 +20,15 @@ object Delete
      * contain all deleted tables.
      * @param deletedTables The tables from which rows are deleted (shouldn't be empty)
      */
-    def apply(target: SqlTarget, deletedTables: Seq[Table]) =
-        SqlSegment(s"DELETE ${ deletedTables.map { _.sqlName }.mkString(", ") } FROM") + target.toSqlSegment
+    def apply(target: SqlTarget, deletedTables: Seq[Table]) = {
+        if (deletedTables.isEmpty)
+            SqlSegment("SELECT NULL FROM") + target.toSqlSegment
+        else
+            SqlSegment(s"DELETE ${ deletedTables.map { _.sqlName }.mkString(", ") } FROM",
+                events = Some(_ => deletedTables.map { table => DataDeleted(table) })) +
+                target.toSqlSegment
+    }
+    
     /**
       * Creates an sql segment that deletes rows from a single table. The deleted
       * table must be included in the provided target.
@@ -38,7 +46,8 @@ object Delete
      * @param table Table being targeted by this delete statement
      * @return A delete statement targeting that table
      */
-    def apply(table: Table) = SqlSegment("DELETE FROM") + table.toSqlSegment
+    def apply(table: Table) = SqlSegment("DELETE FROM", events =
+        Some(_ => Vector(DataDeleted(table)))) + table.toSqlSegment
     
     /**
      * Performs one or more delete queries on a table in a way that deletes only a certain number of items per query.